DisjointSet Union By Size
Union by size:
This is as same as the Union by rank method except this method uses the size to compare the components while connecting. That is why we need a ‘size’ array of size N(no. of nodes) instead of a rank array. The size array will be storing the size for each particular node i.e. size[i] will be the size of the component starting from node i.

Typically, the size of a node refers to the number of nodes that are connected to it.

Algorithm:
Initial configuration:
size array: This array is initialized with one.
parent array: The array is initialized with the value of nodes i.e. parent[i] = i.

The algorithm steps are as follows:
Firstly, the Union function requires two nodes(let’s say u and v) as arguments. Then we will find the ultimate parent (using the findPar() function discussed earlier) of u and v. Let’s consider the ultimate parent of u is pu and the ultimate parent of v is pv.
After that, we will find the size of pu and pv i.e. size[pu] and size[pv].
Finally, we will connect the ultimate parent with a smaller size to the other ultimate parent with a larger size. But if the size of the two is equal, we can connect any parent to the other parent.
While connecting in both cases we will increase the size of the parent node to whom we have connected by the size of the other parent node which is actually connected.